+++ config/src/lua.rs
@@ -254,6 +254,10 @@ pub fn make_lua_context(config_file: &Path) -> anyhow::Result<Lua> {
             "run_child_process",
             lua.create_async_function(run_child_process)?,
         )?;
+        wezterm_mod.set(
+            "background_child_process",
+            lua.create_async_function(background_child_process)?,
+        )?;
         wezterm_mod.set("on", lua.create_function(register_event)?)?;
         wezterm_mod.set("emit", lua.create_async_function(emit_event)?)?;
         wezterm_mod.set("sleep_ms", lua.create_async_function(sleep_ms)?)?;
@@ -806,6 +810,26 @@ async fn run_child_process<'lua>(
         output.stdout.into(),
         output.stderr.into(),
     ))
+}
+
+async fn background_child_process<'lua>(_: &'lua Lua, args: Vec<String>) -> mlua::Result<()> {
+    let mut cmd = smol::process::Command::new(&args[0]);
+
+    if args.len() > 1 {
+        cmd.args(&args[1..]);
+    }
+
+    #[cfg(windows)]
+    {
+        use smol::process::windows::CommandExt;
+        cmd.creation_flags(winapi::um::winbase::CREATE_NO_WINDOW);
+    }
+
+    cmd.stdin(smol::process::Stdio::null())
+        .spawn()
+        .map_err(|e| mlua::Error::external(e))?;
+
+    Ok(())
 }
 
 fn permute_any_mods<'lua>(